src/app/core/interceptors/error-popup.interceptor.ts
import { catchError } from 'rxjs/operators';
import { Injectable } from '@angular/core';
import {
  HttpRequest,
  HttpHandler,
  HttpEvent,
  HttpInterceptor,
} from '@angular/common/http';
import { Observable } from 'rxjs';
import { MatSnackBar } from '@angular/material/snack-bar';

@Injectable()
export class ErrorPopupInterceptor implements HttpInterceptor {
  constructor(private snackBar: MatSnackBar) {}

  intercept(
    request: HttpRequest<unknown>,
    next: HttpHandler
  ): Observable<HttpEvent<unknown>> {
    return next.handle(request).pipe(
      catchError((err) => {
        // the JWT interceptor handles 401 errors
        if (err.status != 401)
          this.snackBar.open(
            'An error has occurred, please try again later',
            undefined,
            {
              duration: 1000,
            }
          );
        throw err;
      })
    );
  }
}
